+++ app/Models/EventTournament.php
@@ -1,6 +1,4 @@
-<?php
-
-namespace App;
+<?phpnamespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
