+++ app/Models/TableProperties.php
@@ -1,7 +1,4 @@
-<?php
-
-
-namespace App;
+<?phpnamespace App\Models;
 
 
 class TableProperties
