+++ app/Models/User.php
@@ -1,6 +1,4 @@
-<?php
-
-namespace App;
+<?phpnamespace App\Models;
 
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
